Use uintptr_t rather than unsigned to hold values that are actually pointers bugid: 6196

+++ Model/ComponentInterface2/CreateComponent.cpp
@@ -3,16 +3,16 @@
 #include "ScienceAPI2.h"
 
 class CMPComponentInterface;
-typedef  unsigned (STDCALL createComponentType)(ScienceAPI2* scienceAPI);
-typedef  void (STDCALL deleteComponentType)(unsigned component);
+typedef  uintptr_t (STDCALL createComponentType)(ScienceAPI2* scienceAPI);
+typedef  void (STDCALL deleteComponentType)(uintptr_t component);
 
-unsigned CreateComponent(ScienceAPI2* scienceAPI, CMPComponentInterface*, const char* dllFileName, void* dllHandle)
+uintptr_t CreateComponent(ScienceAPI2* scienceAPI, CMPComponentInterface*, const char* dllFileName, void* dllHandle)
    {
    createComponentType *createComponent = (createComponentType *) dllProcAddress(dllHandle, "createComponent");
    if (createComponent == NULL) throw std::runtime_error("Missing createComponent() in " + std::string(dllFileName));
-   return (unsigned) createComponent(scienceAPI);
+   return (uintptr_t) createComponent(scienceAPI);
    }
-void DeleteComponent(unsigned component, void* dllHandle)
+void DeleteComponent(uintptr_t component, void* dllHandle)
    {
    deleteComponentType *deleteComponent = (deleteComponentType *) dllProcAddress(dllHandle, "deleteComponent");
    if (deleteComponent == NULL) throw std::runtime_error("Missing deleteComponent()");
